feat(bgp)!: distinct state from enabled intent
Before this commit, both the state and the up/down intent were merged.

It was preventing us from declaring a BGP session in maintenance while
shutting the BGP only on one side.

With this commit, state and enabled are completely separate.

We do not reuse "status" from NetBox DCIM:
- to avoid conflict if we want to redefined the state at the device
  level
- "state" is more appropriate to define if a device is in
  production/validation etc... Status is really accurate.

"enabled" is the same as in OpenConfig and minimize possible
misinterpretation of what it does.